sha256

+++ base64.h
@@ -16,10 +16,10 @@ auto AtoB = [](char* ascii, char*& res) -> size_t {
     memcpy(s, ascii, len);
     res = new char[len * 2];
     size_t cnt = 0;
-    for (int i = 0; i < len; i += 3) {
+    for (size_t i = 0; i < len; i += 3) {
         int group = (s[i] << 16) + (s[i + 1] << 8) + s[i + 2], tcnt = 0;
         char temp[8];
-        for (int j = 0; j < 4; j++, group >>= 6) temp[tcnt++] = group - ((group >> 6) << 6);
+        for (size_t j = 0; j < 4; j++, group >>= 6) temp[tcnt++] = group - ((group >> 6) << 6);
         for (; tcnt; ) res[cnt++] = sheet[temp[--tcnt]];
         res[cnt] = '\0';
     }
@@ -29,13 +29,13 @@ auto BtoA = [](char* base64, char*& res) -> size_t {
     size_t len = strlen(base64);
     char* s = new char[len + 16];
     memset(s + len, 0, sizeof(s));
-    for (int i = 0; i < len; i++)s[i] = reverse_sheet(base64[i]);
+    for (size_t i = 0; i < len; i++)s[i] = reverse_sheet(base64[i]);
     res = new char[len * 2];
     size_t cnt = 0;
-    for (int i = 0; i < len; i += 4) {
+    for (size_t i = 0; i < len; i += 4) {
         size_t group = (s[i] << 18) + (s[i + 1] << 12) + (s[i + 2] << 6) + s[i + 3], tcnt = 0;
         char temp[8];
-        for (int j = 0; j < 3; j++, group >>= 8) temp[tcnt++] = group - ((group >> 8) << 8);
+        for (size_t j = 0; j < 3; j++, group >>= 8) temp[tcnt++] = group - ((group >> 8) << 8);
         for (; tcnt; ) res[cnt++] = temp[--tcnt];
         res[cnt] = '\0';
     }
